increased chunk size warning limit in delta vite build
Former-commit-id: 55cc7c1122164ceddf519e3cd4bc3e3fb04b24bf

+++ delta/src/features/layout/hooks/index.ts
@@ -23,7 +23,7 @@ export type LayoutCreator = (props: LayoutCreatorProps) => Layout;
 
 export type LayoutPlacer = (layout: Layout | LayoutCreator) => void;
 
-export type LayoutRemover = (key: string) => void;
+export type LayoutRemover = () => void;
 
 export const useLayoutPlacer = (): LayoutPlacer => {
   const dispatch = useDispatch();
@@ -49,6 +49,7 @@ export const useLayoutPlacer = (): LayoutPlacer => {
 export const useLayoutRemover = (key: string): LayoutRemover => {
   const dispatch = useDispatch();
   const layout = useSelectLayout(key);
+  if (layout == null) throw new Error(`layout with key ${key} does not exist`);
   return () => {
     dispatch(removeLayout(key));
     if (layout.location === "window") dispatch(closeWindow(key));
